default processing_thread_pool based on
...platform and number of detected processors. On non-GIL
interpreters, default to 1 less than number of processors.
(Should be equal to number of processors instead?)

On MRI, default to 1. I think 1 is better than 0 here, so
if the 'main' thread the Reader is running in is blocked
on I/O, the mapping in it's own 1 thread can still be switched
in to do work. But if someone wants to do some actual tests
in MRI, we could determine for sure if 0 or 1 is best, this is
just my best guess.