DRY source_record_id in Context where it belongs
Tell don't ask, it belongs in Context, and now it's DRY between
indexer and writer (and future needs).

This did require me to rearrange Indexer a bit, creating
the context outside the thread, so it would be available
when I needed it for batch logging. (Alternately batch
logging could have been moved inside the thread?). I think
this is just fine, and even makes the code cleaner, but now
I'm getting paranoid that every change might be a perf regression.
I think it's fine though.