Make writer_class reliable again
It is no longer just a wrapper for settings, as the
actual writer can be set by other means that the writer_class_name
setting.

Setting `indexer.writer_class = SomeClass` still works, with the
tests to prove it. Also test that reader `indexer.writer_class`
still works in all cases.