Allow accumulator to contain a nil
We're currently ignoring 'nil' values (both in logic and by
calling `compact!` on the accumulator), presumably so we
could be sloppy about stuffing stuff into the accumulator.
This breaks the use case (which @jcoyne has) where
nil as a value indicates "delete the current value", while
a missing key says "leave the current value alone". He's got
a SPARQL endpoint, but this would be true for things like
ORM models as well.

The patch and test allows nil values in the accumulator. Our
tests all run fine, but I'm going to have to do more testing to
see if it breaks anyting I currently have in production before
pushing out a release.