Changes to Traject::DebugWriter
Require it by default (#119)

Be more forgiving about the record ID field. Instead
of erroring out when encountering a record without
an 'id' field (or, more generally, whatever is set in
`debug_writer.idfield`, we now throw a warning with
the record number where it first encountered a record
without an identifiable record id in the configured field,
and substitutes a string containing the record number of
the form `record_no_1` as the unique record id.

Tests added.