Be more forgiving about the record ID field
DebugWriter errors out when encountering a record without
an 'id' field (or, more generally, whatever is set in
`debug_writer.idfield`.

This throws a warning with the record number where it
first encouters a record without an identifiable record id
in the configured field, and substitutes a string
containing the record number of the form `record_no_1`
as the unique record id.

Tests added.