NokogiriReader no longer uses streaing parser, too hard to get right
What we had before ended up double-parsing everything, and having correspondingly measured around half the throughput of this one. Even worse on JRuby, because of the way we got around namespace issues in jruby.

This will load the whole input XML into RAM as a DOM doc, but actually performs _better_. I think in most use cases, this will be fine.

A trickier streaming implementation may be possible for best of both worlds, but not that I can get today.