SolrJsonWriter::MaxSkippedRecordsExceeded will now have a #cause
that is the last error encountered that triggered max skipped records. (if max skipped records is 0, as per default, it's the only error)

We have to raise in a rescue to get ruby 2.1+ behavior where thing we rescued is available in #cause, and will be printed by for instance some standard Rails exception printers.

For HTTP response errors, that previously counted as skippable and still do, we have a new error class, Traject::SolrJsonWriter::BadHttpResponse