Allow to not error on 409 solr response status.
Solr can be configured to reject updating a newer document with an older
document bases on a date field. In the case where this happens instead
of updating it's newer local document with the older document, solr will
respond with 409.

Since this can be an expected behavior traject should allow us to
configure it not to throw an error and it should not count these as
skipped towards the calculus of throwing an error when max_skipped is
reached.

This configuration is not set by default.