Allow class-level configure to be called more than once on class.
It was already possible for superclass to have class-level configure without getting clobbered by sub-class. But we need  to be callable more than once in the same specific class -- in part so we can let mixins do it for reusable/shareable code.

So we make the class-level iVar an array of procs instead of a single proc, with every call to  adding to it.

This is adding some more hacks on top of our already somewhat hacky class-level configure impleementation from https://github.com/traject/traject/pull/213 . But it works. May close #237